ci: Corrects fixtures
