+++ pyha_tests/test_pyha.py
@@ -1,35 +1,19 @@
 '''Tests PyHa environment
 '''
-import json
-import os
 from pathlib import Path
-from tempfile import TemporaryDirectory
-from typing import Dict
-import shutil
 import importlib
-import pytest
-from nas_unzip.nas import nas_unzip
 
 from PyHa.IsoAutio import generate_automated_labels
 
 
-@pytest.fixture(name='creds')
-def create_creds() -> Dict[str, str]:
-    """Obtains the credentials
+def test_reference_data(reference_data: Path):
+    """Tests that the reference data is sane
 
-    Returns:
-        Dict[str, str]: Username and password dictionary
+    Args:
+        reference_data (Path): Path to reference data
     """
-    if Path('credentials.json').is_file():
-        with open('credentials.json', 'r', encoding='ascii') as handle:
-            return json.load(handle)
-    else:
-        value = os.environ['NAS_CREDS'].splitlines()
-        assert len(value) == 2
-        return {
-            'username': value[0],
-            'password': value[1]
-        }
+    for i in range(11):
+        assert reference_data.joinpath(f'ScreamingPiha{i + 1}.wav').exists()
 
 def test_resampy_installed():
     try:
@@ -38,57 +22,47 @@ def test_resampy_installed():
     except ImportError:
         assert False, "resampy is not installed"
         
-def test_pyha(creds):
+def test_pyha(reference_data: Path):
     """Tests PyHa
     """
-    with TemporaryDirectory() as tmp_dir:
-        path = Path(tmp_dir).resolve()
-        if not Path('TEST').is_dir():
-            nas_unzip(
-                network_path='smb://e4e-nas.ucsd.edu:6021/temp/github_actions/pyha/pyha_test.zip',
-                output_path=path,
-                username=creds['username'],
-                password=creds['password']
-            )
-        else:
-            shutil.copytree(Path('TEST'), path, dirs_exist_ok=True)
-        isolation_parameters = {
-            "model": "tweetynet",
-            "tweety_output": True,
-            "technique": "steinberg",
-            "threshold_type": "median",
-            "threshold_const": 2.0,
-            "threshold_min": 0.0,
-            "window_size": 2.0,
-            "chunk_size": 5.0,
-            "verbose": True
-        }
-        df = generate_automated_labels(path.as_posix() + '/', isolation_parameters)
-        assert df.empty != True
+    
+    isolation_parameters = {
+        "model": "tweetynet",
+        "tweety_output": True,
+        "technique": "steinberg",
+        "threshold_type": "median",
+        "threshold_const": 2.0,
+        "threshold_min": 0.0,
+        "window_size": 2.0,
+        "chunk_size": 5.0,
+        "verbose": True
+    }
+    df = generate_automated_labels(reference_data.as_posix() + '/', isolation_parameters)
+    assert df.empty != True
 
-        isolation_parameters = {
-            "model": "birdnet",
-            "output_path": "outputs",
-            "lat": 35.4244,
-            "lon": -120.7463,
-            "week": 18,
-            "min_conf": 0.1,
-            "filetype": "wav",
-            "num_predictions": 1,
-            "write_to_csv": False,
-        }
-        df = generate_automated_labels(path.as_posix() + '/', isolation_parameters)
-        assert df.empty != True
+    isolation_parameters = {
+        "model": "birdnet",
+        "output_path": "outputs",
+        "lat": 35.4244,
+        "lon": -120.7463,
+        "week": 18,
+        "min_conf": 0.1,
+        "filetype": "wav",
+        "num_predictions": 1,
+        "write_to_csv": False,
+    }
+    df = generate_automated_labels(reference_data.as_posix() + '/', isolation_parameters)
+    assert df.empty != True
 
-        isolation_parameters = {
-            "model":          "microfaune",
-            "technique":       "steinberg",
-            "threshold_type":  "median",
-            "threshold_const": 2.0,
-            "threshold_min":   0.0,
-            "window_size":     2.0,
-            "chunk_size":      5.0,
-            "verbose":     True
-        }
-        df = generate_automated_labels(path.as_posix() + '/', isolation_parameters)
-        assert df.empty != True 
+    isolation_parameters = {
+        "model":          "microfaune",
+        "technique":       "steinberg",
+        "threshold_type":  "median",
+        "threshold_const": 2.0,
+        "threshold_min":   0.0,
+        "window_size":     2.0,
+        "chunk_size":      5.0,
+        "verbose":     True
+    }
+    df = generate_automated_labels(reference_data.as_posix() + '/', isolation_parameters)
+    assert df.empty != True
